Add HTTP Auth and SSL to the ES output plugin
The typical use case is the placement of a transparent proxy in front of
ES. This PR enables basic Auth and HTTPS to be configured independently.

This patch makes logstash use only elasticsearch-ruby gem when writing
to elasticsearch. Also leverages the manticore http transport that ships
with es-ruby for the basic http auth and ssl features.

It's possible to configure the CA certificate using either a .pem/.cer
file ("cacert" option), or a .jks truststore ("truststore" option).