Improved .travis.yml to support multiple ES versions + use better formatting
The .travis.yml file should support multiple ES versions, this fixes this. It also cleans up the output to be more legible and no longer interleaves the background ES process logs with the rspec logs. It will barf out the background logs in the event of an error. The test script is now external because the .travis.yml file was growing unwieldy.

This also changed the download link which was 2.x only, now we use the universal redirecting link.

Also bumped the JRuby version to 1.7.24

Fixes #387