Remove Unmatched Fields from Template
Currently the template includes specific types, such as "byte" and "float", which are never recognized by
the dynamic template system. The template system instead checks for the widest variant of the data type
(e.g., "long" for integers and "double" for floating point).

This removes those unmatched types to avoid the assumption that they are matched.