Add "version" to 5.x template
This adds a version value of today's date (20160915) as the version. Any change to the template should increase the version number (change to that date), thus making it easier to detect changes.

As this is the first release of 5.x, it really is somewhat moot, but any future release may also choose to build in variable for the 5.x change so that the file only needs to be loaded _if_ the version actually needs to be replaced.