Fix handling of query strings in bulk_path
This allows for query strings in the bulk path to work correctly.
The implementation required extra parsing of the query string to work
correctly.

This also restores some of the slash de-duplication logic removed when
we switched over to java.net.URI compatibility since it simplifies the
logic here.

Fixes #607