Fix escaping behavior and add tests around it (#621)
This fixes #618.

This depends on elastic/logstash#7687 to actually run. Until that is merged and released integration tests here will fail. PR review should be done manually against that branch for both the INTEGRATION and SECURE_INTEGRATION branches for now.

This branch fixes escaping for the user parameter and also adds integration tests for secure scenarios, where user/password are used, including with odd escaped values. This all tests SSL to boot.