Fix support for Elasticsearch 7.x
This commit fixes issues with supporting Elasticsearch 7.x

This handles the removal of the '_routing' field name and
'_retry_on_conflict' update action aliases in place of 'routing' and
'retry_on_conflict' respectively in Elasticsearch 7.x.

Also fixes integration tests to handle these changes and
the renaming of doc type from 'doc' to '_doc' and the mapping
name change from `_default_` to '_doc'.