Look up cluster_uuid
This change allows the plugin to inform other parts of Logstash about
which ES cluster it is connecting to.

It is built on top of the Plugin Metadata registry: https://github.com/elastic/logstash/pull/10691

Note:
The DLQ tests were written to just assume a clean buffer in the logs
but this is not a safe assumption.

Now we temporarily replace the logging subsystem with a double
during these tests so that we can be assured that we are checking
a clean instance.