Clarify load balance behavior if array of hosts is used
I've seen the question, on how the output load balances, coming up every now and then and the docs should have the answer.
I'd still like a logstash developer to review and double check that it still is Round-robin per bulk in the current version of the plugin. 
My assumption is based on https://github.com/logstash-plugins/logstash-output-elasticsearch/issues/274 and https://github.com/logstash-plugins/logstash-output-elasticsearch/issues/493#issuecomment-389808368, but the first one is from 2015 and the second onw is from 2018, but with Logstash 2.4.