fix breakage caused by external use of our internal API
Because Logstash Core v7.7.0-v7.8.x override the `use_event_type?` in a
monitoring implelentation's subclass with a required (and ignored) `client`
argument, when we send `use_event_type?` internally, we must send an argument
or else we break compatibility with those versions of Logstash.

This change re-adds the required argument and modifies our call-site to send
with an arity of 1, re-names the resulting local variable to be descriptive
of our intent, and adds commentary to prevent future accidental breakage.