fix: prevent sub-batch 413's from blocking whole batch (#972)
* fix: prevent sub-batch 413's from infinitely retrying whole batch

The Http Client breaks batches of actions into sub-batches that are up to 20MB
in size, sending larger actions as batches-of-one, and zips the responses
together to emulate a single batch response from the Elasticsearch API.

When an individual HTTP request is rejected by Elasticsearch (or by an
intermediate proxy or load-balancer) with an HTTP/1.1 413, we can emulate
the error response instead of blowing an exception through to the whole batch.
This allows only the offending events/actions to be subject to retry logic.

Along the way, we improve logging at the `debug` level for sub-batches, and
emit clear `warn`-level logs with payload sizes when we hit HTTP 413 rejections.

* size batch by _decompressed_ payload size

* tests: config elasticsearch to allow wildcard deletes

The default value of Elasticsearch's `action.destructive_requires_rename` has
changed to true in https://github.com/elastic/elasticsearch/pull/66908 which
causes our integration tests' wildcard deletes to fail. By specifying this
config explicitly, we ensure the desired behaviour is selected.