data-streams: ECS is pre-req for `data_streams: true`
Prevent the plugin from starting up unless it is in a non-disabled ECS mode.

Running this plugin with `ecs_compatibility: disabled` is a recipe for runtime
mapping conflicts, since it usually indicates that the _pipeline_ is not being
run in ECS-compatibility mode, which means nearly every other upstream plugin
in the pipeline will add metadata in non-ECS-friendly way.