Log bulk request response body on error, not just when debug logging is set
Prior to this commit, an error response to a bulk request would only log the
response body when debug logging is enabled. This removes an important piece
of information when tracking down issues, which is particularly important in
this case, as an error response to a bulk request may result in requests being
retried indefinitely, causing pipeline stalls.

This commit adds the response body to the `log_hash` regardless of the log
level,so 429 responses will remain unchanged, and will only log when debug is
enabled, but other error responses will include the response body to add this
additional context to help track down issues

Relates: #999