Fixed abort condition in case of shutdown is requested during connection wait.
While waiting for connection to be established, the busy wait loop that also check for shutdown request.
The abort of the batch should happen if no connection was successful established and there is a shutdown request;
else it could be that the loop exit condition is trigger with both parts satisfied and it's not correct to abort the batch if a connection was created.