Cleanup ci refresh rollout
This commit updates the `docker-setup.sh` and `.travis.yml` files to follow the
current patterns in the CI refresh rolled out to accomidate the new release
cadence. The specific changes are explicitly passing a parameter to the docker
pull function in the setup script and ensuring that the travis config segments
steps correctly.