Properly handle 413 Payload Too Large errors (#1199)
* Properly handle `413` Payload Too Large errors

Previously when Elasticsearch responds with a 413 (Payload Too Large) status,
the manticore adapter raises an error before the response can be processed
by the bulk_send error handling. This commit refactors the way
`BadErrorResponse` codes are handled. Previously we had logic in the manticore
adaptor which special cased raising errors on some codes. This commit refactors
such that the adaptor raises on any error status and the caller is now
responsible for special case handling the code.

* 12.0.2 release prep

* Use `error_code` instead of `code` when handling BadResponseCodeError

Previously a few bugs spotted in code review were being obfuscated by the
combinations of tests not running in CI and the incorrect method for retrieving
a code from a BadResponseCodeError. This commit updates the method names and
addresses the feedback from code review.

---------

Co-authored-by: João Duarte <jsvduarte@gmail.com>