Track logstash versions in .ci (#1221)
* Track logstash versions in .ci

Previously logstash versions were tracked in the logstash repo and curled during
CI steps. This commit moves the source of truth to the `.ci` repo where the
consumer of that information lives. This also changes the format from json to
yaml to allow commenting in the file.

* REMOVE THIS COMMIT

Temporarily include logstash-versions until https://github.com/logstash-plugins/.ci/pull/77
is merged.

* Fix path to logstash-releases.yaml

* Code review resolution

* Revert "REMOVE THIS COMMIT"

This reverts commit 2aadedabb539e0ecd23d074e2c9cb082ca7c7a4f.