Use docker-setup.sh from upstream .ci repo
The custom docker-setup.sh script here is not materially different from the
shared script in the `.ci` repository. The only difference were pulling and
starting the elasticsearch image. This should not affect CI materially and the
consistency gain we get using the same `docker-setup.sh` file everywhere is
preferrable to having a slightly different script here.